+++ releasenotes.md
@@ -1,6 +1,7 @@
 This release includes:
 ----------------------
-Various bug fixes
+Changed how the % limit is been shown in ingredients page
 Versioning format changed from x.x.x to x.x
 Added release notes
 IFRA certification improvements
+Various bug fixes